Improved the Mea Solution Folder Structure
