--- Kmd.Momentum.Mea/MeaServiceCollectionExtensions.cs
@@ -1,12 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Kmd.Momentum.Mea
-{
-    public static class MeaServiceCollectionExtensions
-    {
-        public static IServiceCollection AddMea(this IServiceCollection services)
-        {
-            return services;
-        }
-    }
-}
