+++ kmd-momentum-mea.sln
@@ -18,9 +18,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kmd.Momentum.Mea.Integration.Tests", "test\Kmd.Momentum.Mea.Integration.Tests\Kmd.Momentum.Mea.Integration.Tests.csproj", "{BE9F3981-F7D7-4231-8227-C8D06EC7DE62}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kmd.Momentum.Mea.Common", "Kmd.Momentum.Mea.Common\Kmd.Momentum.Mea.Common.csproj", "{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kmd.Momentum.Mea", "src\Kmd.Momentum.Mea\Kmd.Momentum.Mea.csproj", "{240204DA-B608-4B67-880D-3883CA3FB7AC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kmd.Momentum.Mea", "Kmd.Momentum.Mea\Kmd.Momentum.Mea.csproj", "{5988960B-639D-4671-ACFE-1D18197EC625}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kmd.Momentum.Mea.Common", "src\Kmd.Momentum.Mea.Common\Kmd.Momentum.Mea.Common.csproj", "{C5CBF257-B009-4EF1-9342-2DC31173E575}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -68,30 +68,30 @@ Global
 		{BE9F3981-F7D7-4231-8227-C8D06EC7DE62}.Release|x64.Build.0 = Release|Any CPU
 		{BE9F3981-F7D7-4231-8227-C8D06EC7DE62}.Release|x86.ActiveCfg = Release|Any CPU
 		{BE9F3981-F7D7-4231-8227-C8D06EC7DE62}.Release|x86.Build.0 = Release|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Debug|x64.Build.0 = Debug|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Debug|x86.Build.0 = Debug|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Release|x64.ActiveCfg = Release|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Release|x64.Build.0 = Release|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Release|x86.ActiveCfg = Release|Any CPU
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6}.Release|x86.Build.0 = Release|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Debug|x64.Build.0 = Debug|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Debug|x86.Build.0 = Debug|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Release|x64.ActiveCfg = Release|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Release|x64.Build.0 = Release|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Release|x86.ActiveCfg = Release|Any CPU
-		{5988960B-639D-4671-ACFE-1D18197EC625}.Release|x86.Build.0 = Release|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Debug|x64.Build.0 = Debug|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Debug|x86.Build.0 = Debug|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Release|x64.ActiveCfg = Release|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Release|x64.Build.0 = Release|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Release|x86.ActiveCfg = Release|Any CPU
+		{240204DA-B608-4B67-880D-3883CA3FB7AC}.Release|x86.Build.0 = Release|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Debug|x64.Build.0 = Debug|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Debug|x86.Build.0 = Debug|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Release|x64.ActiveCfg = Release|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Release|x64.Build.0 = Release|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Release|x86.ActiveCfg = Release|Any CPU
+		{C5CBF257-B009-4EF1-9342-2DC31173E575}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -100,8 +100,8 @@ Global
 		{54FCEF30-EBB8-4B71-9C60-42BB076BC0B8} = {E1A04C16-5A97-440D-AECD-450348113EC0}
 		{EA12E001-0245-4D84-8CC8-8B4F3EB4CF98} = {0E63EA42-074D-441C-8515-7979BD51991F}
 		{BE9F3981-F7D7-4231-8227-C8D06EC7DE62} = {0E63EA42-074D-441C-8515-7979BD51991F}
-		{CDF37C87-C49F-4D8D-B214-283A5FC81CF6} = {E1A04C16-5A97-440D-AECD-450348113EC0}
-		{5988960B-639D-4671-ACFE-1D18197EC625} = {E1A04C16-5A97-440D-AECD-450348113EC0}
+		{240204DA-B608-4B67-880D-3883CA3FB7AC} = {E1A04C16-5A97-440D-AECD-450348113EC0}
+		{C5CBF257-B009-4EF1-9342-2DC31173E575} = {E1A04C16-5A97-440D-AECD-450348113EC0}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {AFE37BFF-F48B-489F-A9ED-8D782D46C1C7}
