+++ src/Kmd.Momentum.Mea.Common/Authorization/HasResourceHandler.cs
