+++ src/Kmd.Momentum.Mea.Common/Authorization/HasResourceRequirement.cs
