+++ src/Kmd.Momentum.Mea.Common/Authorization/Resource.cs
