+++ src/Kmd.Momentum.Mea/MeaServiceCollectionExtensions.cs
@@ -1,4 +1,7 @@
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Collections.Generic;
+using System.Text;
 
 namespace Kmd.Momentum.Mea
 {
